move activity content into own folder
closes #323
