+++ projects/ngx-d3timeline/src/lib/activity/activity.ts
