+++ projects/ngx-d3timeline/src/lib/drag-utils.ts
@@ -1,7 +1,7 @@
 import { Orientation } from './core/orientation';
 import { TimeScale, InverseBandScale } from './scale-types';
 import { TimelineDragEvent } from './content/timeline-drag-event';
-import { PositionedActivity } from './positioned-activity';
+import { PositionedActivity } from './activity/positioned-activity';
 import { identifier } from './core/types';
 import { Point, origin } from './core/point';
