+++ projects/ngx-d3timeline/src/lib/scale-utils.ts
@@ -1,7 +1,7 @@
 import { TimeScale, BandScale, InverseBandScale } from './scale-types';
 import { Orientation } from './core/orientation';
 import { scaleBand, scaleTime } from 'd3-scale';
-import { Activity } from './activity';
+import { Activity } from './activity/activity';
 import { min, max } from 'd3-array';
 import { View } from './view/view';
